alter metadata sizes and add methods to create entries

+++ src/char_code.rs
@@ -2,11 +2,11 @@
 pub struct CharCodePair {
   pub value: char,
   pub bits: u8,
-  pub code: u8
+  pub code: u32,
 }
 
 impl CharCodePair {
-  pub fn new(value: char, bits: u8, code: u8) -> CharCodePair {
+  pub fn new(value: char, bits: u8, code: u32) -> CharCodePair {
     return CharCodePair {
       value,
       bits,
